fix: revert the `--tests` flag behavior
Fixes #297

To avoid confusions, the `--tests` flag should include exact *one* set
of test configurations. In the previous versions, it was Vitest for unit
testing and Cypress for E2E testing.
So for consistency, we remove the newly added Nightwatch template from
it.

But as we have more and more testing options, I would encourage everyone
to specify the name of the testing framework they want to use, rather
than use the general `--tests` flag. It would also be less prone to bugs